+++ dallinger/frontend/static/scripts/dallinger2.js
@@ -428,6 +428,38 @@ var dallinger = (function () {
   };
 
   /**
+   * Load an existing `Participant` into the dlgr.identity by making a ``POST``
+   * request to the experiment `/participant` route with an ``assignment_id``.
+   *
+   * @returns {jQuery.Deferred} See :ref:`deferreds-label`
+   */
+  dlgr.loadParticipant = function(assignment_id) {
+    var deferred = $.Deferred(),
+        url = '/participant';
+
+    if (dlgr.identity.participantId !== undefined && dlgr.identity.participantId !== 'undefined') {
+      deferred.resolve();
+    } else {
+      $(function () {
+        $('.btn-success').prop('disabled', true);
+        dlgr.post(url, {assignment_id: assignment_id}).done(function (resp) {
+          console.log(resp);
+          $('.btn-success').prop('disabled', false);
+          dlgr.identity.participantId = resp.participant.id;
+          dlgr.identity.recruiter = resp.participant.recruiter_id;
+          dlgr.identity.hitId = resp.participant.hit_id;
+          dlgr.identity.workerId = resp.participant.worker_id;
+          dlgr.identity.assignmentId = assignment_id;
+          dlgr.identity.mode = resp.participant.mode;
+          dlgr.identity.fingerprintHash = resp.participant.fingerprint_hash;
+          deferred.resolve();
+        });
+      });
+    }
+    return deferred;
+  };
+
+  /**
    * Creates a new experiment `Node` for the current partcipant.
    *
    * @example